requirements.txt
blinker==1.9.0
bottle==0.13.4
click==8.3.1
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
proxy_tools==0.1.0
pyobjc-core==12.1
pyobjc-framework-Cocoa==12.1
pyobjc-framework-Quartz==12.1
pyobjc-framework-Security==12.1
pyobjc-framework-UniformTypeIdentifiers==12.1
pyobjc-framework-WebKit==12.1
pywebview==6.1
typing_extensions==4.15.0
Werkzeug==3.1.5
